Removing explicit net46 build targets
Also changed IGameState to IQuestStates and made extension method for GetPotency return a float not an int.

+++ src/OpenRpg.Core/Extensions/EffectExtensions.cs
@@ -6,10 +6,10 @@ namespace OpenRpg.Core.Extensions
 {
     public static class EffectExtensions
     {
-        public static int GetPotencyFor(this IEnumerable<Effect> effects, int effectType)
-        { return (int)effects.Where(x => x.EffectType == effectType).Sum(x => x.Potency); }
+        public static float GetPotencyFor(this IEnumerable<Effect> effects, int effectType)
+        { return effects.Where(x => x.EffectType == effectType).Sum(x => x.Potency); }
         
-        public static int GetPotencyFor(this IEnumerable<Effect> effects, params int[] effectTypes)
-        { return (int)effects.Where(x => effectTypes.Contains(x.EffectType)).Sum(x => x.Potency); }
+        public static float GetPotencyFor(this IEnumerable<Effect> effects, params int[] effectTypes)
+        { return effects.Where(x => effectTypes.Contains(x.EffectType)).Sum(x => x.Potency); }
     }
 }
